requirements.txt
frappe
python-gcm